refactor. StoreImg to Image
