+++ dishcovery/src/main/java/capstone/miso/dishcovery/domain/image/Image.java
@@ -1,4 +1,4 @@
-package capstone.miso.dishcovery.domain.storeimg;
+package capstone.miso.dishcovery.domain.image;
 
 /**
  * author        : duckbill413
@@ -18,7 +18,7 @@ import lombok.*;
 @Entity
 @ToString(callSuper = true)
 @EqualsAndHashCode(callSuper = true)
-public class StoreImg extends BaseEntity {
+public class Image extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long sid;
