+++ dishcovery/src/main/java/capstone/miso/dishcovery/domain/store/dto/StoreDetailDTO.java
@@ -26,7 +26,7 @@ public class StoreDetailDTO {
         private String address;
         private String category;
         private String sector;
-        private String mainImageUrl;
+        private String mainImage;
         private List<String> onInfo;
         private List<String> offInfo;
         private List<String> keywords;
