--- dishcovery/src/main/java/capstone/miso/dishcovery/domain/storeimg/dto/StoreImageDto.java
@@ -1,16 +1,0 @@
-package capstone.miso.dishcovery.domain.storeimg.dto;
-
-import jakarta.validation.Valid;
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * author        : duckbill413
- * date          : 2023-04-14
- * description   :
- **/
-@Getter
-@Setter
-public class StoreImageDto {
-    private PhotoViewer photoViewer;
-}
