--- dishcovery/src/main/java/capstone/miso/dishcovery/domain/storeimg/repository/StoreImgRepository.java
@@ -1,13 +1,0 @@
-package capstone.miso.dishcovery.domain.storeimg.repository;
-
-import capstone.miso.dishcovery.domain.storeimg.StoreImg;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * author        : duckbill413
- * date          : 2023-04-14
- * description   :
- **/
-
-public interface StoreImgRepository extends JpaRepository<StoreImg, Long> {
-}
