+++ dishcovery/src/test/java/capstone/miso/dishcovery/domain/image/service/StoreImgServiceTest.java
@@ -1,13 +1,8 @@
-package capstone.miso.dishcovery.domain.storeimg.service;
+package capstone.miso.dishcovery.domain.image.service;
 
-import capstone.miso.dishcovery.domain.store.Store;
 import capstone.miso.dishcovery.domain.store.repository.StoreRepository;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
 
 /**
  * author        : duckbill413
